Add the Helm default variable of hostingClusterCapabilities (#149)
When deploying in hosted mode, the managed cluster's Kubernetes version
is always used in Helm charts conditionals such as:
{{- if semverCompare "< 1.16.0" .Capabilities.KubeVersion.Version }}

The hosted mode's version is required to guard against using Kubernetes API
features that aren't present on the cluster that the addon is being
deployed to not necessarily based on the target cluster that addon will
manage.

The new default Helm chart variable of hostingClusterCapabilities
is added so that in hosted mode, the hosted cluster's Kubernetes
version Kubernetes version is available. Unless the AgentAddonFactory
was built with the WithHostingCluster method, the
hostingClusterCapabilities will have empty values in it. The empty
values being returned by default make it easier to make if statements in
the Helm chart rather than guarding against a value not being present.

Relates:
https://issues.redhat.com/browse/ACM-3233
https://issues.redhat.com/browse/ACM-2923

Signed-off-by: mprahl <mprahl@users.noreply.github.com>