Fix integration test expectation for dynamic aggregate ClusterRole
Update test expectation from 46 to 47 related resources to account for the
dynamically created aggregate ClusterRole when addon-manager is enabled.

The test 'Deployment should be updated when clustermanager is changed' was
failing because our dynamic aggregate ClusterRole implementation correctly
creates an additional resource that wasn't accounted for in the test.

Generated by Augment.

Signed-off-by: zhujian <jiazhu@redhat.com>